Selftests: split functional tests for utils modules
The selftests/functional/test_utils.py file contains tests for both
the "avocado.utils.process" and "avocado.utils.filelock" module.

This splits it, moving the new files into the "utils" subdir.
