+++ selftests/functional/utils/test_process.py
@@ -1,14 +1,9 @@
-import multiprocessing
 import os
-import random
 import stat
 import sys
 import time
-import unittest
 
 from avocado.utils import process
-from avocado.utils.filelock import FileLock
-from avocado.utils.stacktrace import prepare_exc_info
 from selftests.utils import TestCaseTmpDir, skipOnLevelsInferiorThan
 
 # What is commonly known as "0775" or "u=rwx,g=rwx,o=rx"
@@ -184,38 +179,3 @@ class ProcessTest(TestCaseTmpDir):
         result = proc.run()
         self.assertEqual(result.exit_status, 0, f"result: {result}")
         self.assertIn(b"load average", result.stdout)
-
-
-def file_lock_action(args):
-    path, players, max_individual_timeout = args
-    max_timeout = max_individual_timeout * players
-    with FileLock(path, max_timeout):
-        sleeptime = random.random() / 100
-        time.sleep(sleeptime)
-
-
-class FileLockTest(TestCaseTmpDir):
-    @skipOnLevelsInferiorThan(3)
-    def test_filelock(self):
-        """
-        :avocado: tags=parallel:1
-        """
-        # Calculate the timeout
-        start = time.monotonic()
-        for _ in range(50):
-            with FileLock(self.tmpdir.name):
-                pass
-        timeout = 0.02 + (time.monotonic() - start)
-        players = 500
-        pool = multiprocessing.Pool(players)
-        args = [(self.tmpdir.name, players, timeout)] * players
-        try:
-            pool.map(file_lock_action, args)
-        except Exception:
-            msg = "Failed to run FileLock with %s players:\n%s"
-            msg %= (players, prepare_exc_info(sys.exc_info()))
-            self.fail(msg)
-
-
-if __name__ == "__main__":
-    unittest.main()
